+++ apps/client/src/app/components/symbol-icon/symbol-icon.component.ts
@@ -2,8 +2,9 @@ import {
   ChangeDetectionStrategy,
   Component,
   Input,
-  OnInit
+  OnChanges
 } from '@angular/core';
+import { DataSource } from '@prisma/client';
 
 @Component({
   selector: 'gf-symbol-icon',
@@ -11,12 +12,22 @@ import {
   templateUrl: './symbol-icon.component.html',
   styleUrls: ['./symbol-icon.component.scss']
 })
-export class SymbolIconComponent implements OnInit {
+export class SymbolIconComponent implements OnChanges {
+  @Input() dataSource: DataSource;
   @Input() size: 'large';
+  @Input() symbol: string;
   @Input() tooltip: string;
   @Input() url: string;
 
+  public src: string;
+
   public constructor() {}
 
-  public ngOnInit() {}
+  public ngOnChanges() {
+    if (this.dataSource && this.symbol) {
+      this.src = `../api/v1/logo/${this.dataSource}/${this.symbol}`;
+    } else if (this.url) {
+      this.src = `../api/v1/logo?url=${this.url}`;
+    }
+  }
 }