+++ libs/common/src/lib/interfaces/index.ts
@@ -19,6 +19,7 @@ import { InfoItem } from './info-item.interface';
 import { LineChartItem } from './line-chart-item.interface';
 import { PortfolioChart } from './portfolio-chart.interface';
 import { PortfolioDetails } from './portfolio-details.interface';
+import { PortfolioDividends } from './portfolio-dividends.interface';
 import { PortfolioInvestments } from './portfolio-investments.interface';
 import { PortfolioItem } from './portfolio-item.interface';
 import { PortfolioOverview } from './portfolio-overview.interface';
@@ -31,6 +32,7 @@ import { PortfolioSummary } from './portfolio-summary.interface';
 import { Position } from './position.interface';
 import { BenchmarkResponse } from './responses/benchmark-response.interface';
 import { ResponseError } from './responses/errors.interface';
+import { ImportResponse } from './responses/import-response.interface';
 import { OAuthResponse } from './responses/oauth-response.interface';
 import { PortfolioPerformanceResponse } from './responses/portfolio-performance-response.interface';
 import { ScraperConfiguration } from './scraper-configuration.interface';
@@ -57,11 +59,13 @@ export {
   Filter,
   FilterGroup,
   HistoricalDataItem,
+  ImportResponse,
   InfoItem,
   LineChartItem,
   OAuthResponse,
   PortfolioChart,
   PortfolioDetails,
+  PortfolioDividends,
   PortfolioInvestments,
   PortfolioItem,
   PortfolioOverview,