+++ libs/common/src/lib/interfaces/portfolio-public-details.interface.ts
@@ -9,6 +9,8 @@ export interface PortfolioPublicDetails {
       | 'allocationCurrent'
       | 'countries'
       | 'currency'
+      | 'dataSource'
+      | 'dateOfFirstActivity'
       | 'markets'
       | 'name'
       | 'netPerformancePercent'