+++ libs/common/src/lib/interfaces/responses/import-response.interface.ts
@@ -1,0 +1,5 @@
+import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interface';
+
+export interface ImportResponse {
+  activities: Activity[];
+}