+++ libs/ui/src/lib/holdings-table/holdings-table.module.ts
@@ -8,17 +8,17 @@ import { MatSortModule } from '@angular/material/sort';
 import { MatTableModule } from '@angular/material/table';
 import { RouterModule } from '@angular/router';
 import { GfPositionDetailDialogModule } from '@ghostfolio/client/components/position/position-detail-dialog/position-detail-dialog.module';
+import { GfSymbolIconModule } from '@ghostfolio/client/components/symbol-icon/symbol-icon.module';
 import { GfSymbolModule } from '@ghostfolio/client/pipes/symbol/symbol.module';
 import { GfNoTransactionsInfoModule } from '@ghostfolio/ui/no-transactions-info';
 import { GfValueModule } from '@ghostfolio/ui/value';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 
-import { GfSymbolIconModule } from '../symbol-icon/symbol-icon.module';
-import { PositionsTableComponent } from './positions-table.component';
+import { HoldingsTableComponent } from './holdings-table.component';
 
 @NgModule({
-  declarations: [PositionsTableComponent],
-  exports: [PositionsTableComponent],
+  declarations: [HoldingsTableComponent],
+  exports: [HoldingsTableComponent],
   imports: [
     CommonModule,
     GfNoTransactionsInfoModule,
@@ -37,4 +37,4 @@ import { PositionsTableComponent } from './positions-table.component';
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
-export class GfPositionsTableModule {}
+export class GfHoldingsTableModule {}